updated tiddlerOrigin macro.
last change brought in two new states unsyncedPrivate and unsyncedPublic.
By default the tiddlerOrigin macro would look for tiddlers called
unsyncedPrivateIcon or unsyncedPublicIcon respectively to be rendered.

This change allows you to display the same icon for both  if either are absence
a  "unsyncedIcon" is used instead. I'm unsure whether we want 2 icons for the two
states and would like to think about it more deeply first.

I also feel we might be able to refactor svg icons using the use element
see this post about the use element on this subject:
http://groups.google.com/group/tiddlyweb/browse_thread/thread/4d7847f5cc379dfc/036494d0a3419e44?show_docid=036494d0a3419e44