Altered backstage to work offline and vizualise overall sync status
* Updated status as it is used by the message macro in backstage and was
throwing errors.
* content of user and space tasks now determined by whether
window.location.protocol is file or not. Depending on the circumstances
the different menus in Backstage are showed.
* Took away user lookup in backstage.show as it is no longer important in
the offline mode. "You are unplugged" now shown in the middle menu instead
of you are a member/non-member.
* Changes to support the logo for the middle button to be refreshed and to show
the unsyncedIcon when there are unsynced changes. The list of unsynced tiddlers is
also given class for hiding purposes if empty.
* updated stylesheet to make use of above class
* server side saving plugin updated to refresh the logo
* slight improvements to readability of code.