Added policies to make_fake_space and updated tests to expect this
Some of the tests in test_jsonp.py require the policy set properly on the bags/
recipes. The most obvious way of doing thhis was to update make_fake_space to
set the policies of each bag/recipe correctly. Doing this broke some other tests
though (ie tests that didn't expect a policy on the private bag), so I've also
updated them to make their requests as a logged in user as well.