added some tests to test the root domain as well as subdomains
As it's jsonp, we need to check the root domain as well as subdomains.
Previously, this was safe and being ignored, but with jsonp, its no longer
safe, so we need to DropPrivs for everything whenever jsonp_callback
is specified.

This resulted in a small bit of refactoring on DropPrivs and the introduction
of a new method "_handle_jsonp" which _always_ gets called. This should result
in the desired behaviour (dropping privs) whatever url is called.