remove deleted draft locally. when mint a draft open original and draft
when creating a draft the ui suggests the tiddler has been overwritten.
This change opens both the draft and the original to make it clear this isn't
the case.

this addresses 455 by controversially using store.deleteTiddler rather
than store.removeTiddler. this is because removeTiddler doesn't have a
callback and in the case of a publish we want to refresh the tiddler in
story and run a custom callback. Hoping adaptor rewrite will make this
code cleaner.