minor tweaks
jQuery was throwing exception for titles with @ symbol in
(I believe @ is a reserved character in selectors) thus removed it
to be safe.

Moved displayTiddler to bottom of function as precaution,
removed animate due to exception being thrown during the
displayTiddler function in preso (tobe investigated)
which was preventing important diffs from loading.