autosave by default when offline
currently the save changes button is not exposed anywhere in the backstage UI.
Thus it is essential autoSave is on by default.
Note users can still turn this off and add a savechanges button.
We might want to consider adding these to the backstage space menu at a later date.