allow saving of imported tiddlers
The ImportExternalLinksPlugin
http://following.tiddlyspace.com/#ImportExternalLinksPlugin

enhances the cloneTiddler command to allow cloning of sucked in
tiddlers. This is needed in the ILGA tiddlyspace where articles
from spaces can be cloned and pushed on to the website.

We may want to fold this change to cloneTiddler into the core
however this change provides a hook to allow it to work in the
meantime.