always set secondary_tiddlyweb_user with openid login
This sets a secondary_tiddlyweb_user on the domain *.domain.com for
an openid login. This means once logged in with openid a user can
log in via cookie id and then make use of the secondary tiddlyweb user
for associating that identity.

This will allow plugins to provide an openid experience of not requiring
a password.
e.g. http://space-install.tiddlyspace.com:80/