fixed openid bug
If a user is logged in via their username and then logs in via an
openid and makes use of the hash in the redirect this is a special
case where we don't actually want to log the user in - we only want
to set the secondary_cookie.

When a user logs in via openid normally we do want to set it.

This makes use of a flag to ensure openid adding works.