refactored groupBy plugin to support templating
As we go forward it will be useful to template groupBy results
There are uses for this with sucked in tiddlers where server.title
is different from local title.

this is consistent with http://trac.tiddlywiki.org/ticket/1413
suggesting parameters template and groupTemplate for templating the
outputs of the plugin.