Added username to the csrf cookie.
This fixes a bug whereby a user logging out and then logging in again as
someone else does not get their csrf cookie refreshed.

The fix implemented is to extend the cookie, so that it is now made up
of the following:

	timestamp:username:hash

where the hash is the same as before. This means that the server can check
to see if the user has changed, and refresh the cookie as appropriate.