Initial support for webfinger and profiles in tiddlyspace.
This is part of a larger plan to implement OStatus in tiddlyspace,
incuding webfinger, profiles, activity streams, pubsubhub, etc.

generally following this guide:

  http://ostatus.org/2010/10/04/how-ostatus-enable-your-application

This commit adds

  /.well-known/host-meta

as the discovery mechanism for

 /webfinger?q={uri}

which produces the xrd that points to some resources, including

 /profiles/{username}

Which presents the contents of a [[profile]] tiddler from the user's
public bag (if it is exists) and a list of last 20 tiddlers they have
modified. It's ugly at the moment. Basically just a placeholder.

Some questions to wonder about:

* should the profile have other reprsentations (at the moment is
  points to a search URI for the feed associated with it)
* what are the other links that ought to be in the finger data
* what are the presentations that should be in the profile (e.g.
  hcard)
* how to make the profile "nice". I think it is _critical_ that it
  not be a tiddlywiki as it needs to be an accessible piece of data.
  That doesn't mean it can't be edited or managed by a TiddlyWiki.

Longer term things:

* how to send pubsubhub pings (and where). probably twp.dispatcher
* salmon

This is all a big experiment (part of the reason why it is on a branch)
so nothing is set in stone. Investigating the options.