prevent multiple calls to followTiddler
If a tiddler containing the followTiddler macro is refreshed by
another macro or by the TiddlyWiki refresh mechanism it will trigger
a search for every refresh. This logic makes sure only the last request
on the given tiddler is run meaning only one ajax request will be made

this addresses the problem described in 609.