Deal with missing and shadow tiddler modifiers
Missing and shadow tiddlers have the modifier (missing) and
(built-in shadow tiddler) respectively by default.

This change when a call to renderAvatar is made with these values it will
draw missingIcon or shadowIcon respectively if available. If these icons
do not exist then nothing will be rendered.

Note one limitation that is under our control:
if config.views.wikified.shadowModifier is changed to tiddlywiki for
example this would prevent the tiddlywiki SiteIcon from being displayed.