made filter parameter work
In tsScan a filter parameter was added with the intention of allowing
client side filters to be run on the resulting set of tiddlers. This has
benefits where a client side filter can be run on something that may be
expensive to do on the server.

Since client side filters are extensible this opens up a lot of possibilities
for instance http://spaces.tiddlyspace.com/ uses a custom startsWith filter
to only show tiddlers beginning with a certain letter.