improve how tsScan caches results
the caching mechanism for tsScan is not as good as it could be.
Currently the caching only kicks in after 1 single successful request.
so say I tsScan the url /search?q=title:Foo and that doesn't complete and
then I run another tsScan macro on the same url, another ajax request
will be sent.

To get round this tsScan now queues outstanding callbacks. The request only
happens once and any outstanding callbacks are fired when it finally
returns.