Update profiles so that it uses twp.templates.
This is the start of making use of templates in various ways in
TiddlySpace.

The hope here is that whereas the templates will be quite static,
the CSS that gets used can be overriden in some way by creating
something in a bag related to the thing. For instance a
/bag/user_public/tiddlers/profile.css could override the default
from the common bag.

So one next would be to make that work.

Another next would be to switch the HTML serialization to use
templates for presentation of the usual HTML URIs. Once that's working
options for how to do customizations and controls can be considered.

This is just a way to get the ball rolling.