Add HTML templates for a single tiddler and any collection of tiddlers.
list_tiddlers does some tricks to make sure that links are satisfactory
whether we are in a search, a recipe, a bag or revisions as the container.

One change from the defaults in TiddlyWeb is that when a bag is listed
is it _always_ listed by its bag url, as that's the canonical one and
exposing recipes urls just seems to cause confusion. We may need to
revisit this thinking.

On the tiddler template, the title of the tiddler is now the link
back to the tiddler in space, rather than having a "foo in space"
link.

For now the profiles.css file is being used as a default CSS as I
found the structure of the HTML on the profiles useful and
non-distracting and that CSS works for that.

Since creating these things is explicitly to make it easier for
other people to adapt the HTML and the CSS, starting from a boring
point seems reasonable.

Will need to add templates for /bags, /recipes, /bags/{bag_name and
/recipes/{recipe_name}

Once that's done, it will be time to refactor out the duplication.