expose fields in definition list
TODO: this expects fields as an array. Currently it is a string
so instead not working correctly.

It is expected that some fields will be used to express certain
information. eg. geo.lat and geo.long are used in TiddlySpace for
storing geo information. By recognising this and promoting that information
e.g. linking to a map, it encourages their use more.