recognised geo.lat and geo.long as special fields
this implements the microformat
http://microformats.org/wiki/geo
and recognises fields with these names to be special.
which is a good thing as it promotes their adoption.
we currently have lots of geotiddlers in tiddlyspace due to openbritain.

i've linked the data to open street map to make it useful and to promote
another open source project

with browser plugins such as firefox operator you can export the geographical
information to kml, export to yahoo/google maps etc..