Added csrf handling to the default HTML challengers
Added render_form to the existing openid challenger, and created an
additional cookie_form challenger. Also extracted the getCSRFToken
function into its own tiddler, so that it can be used with other,
non-TiddlyWiki pages.

The CSRF prevention mechanism relies on JavaScript being required to
successfully POST to the server. This means that the HTML challengers
now require JavaScript to successfully log in if the user has already
logged in.