Add experimental support for space-based css and javascript overrides in HTML representations.
If HtmlCss or HtmlJavascript tiddlers exist in the current space's recipes
then:

* the css link in the base template becomes a link to HtmlCss, instead
  of profile.css
* a <script> tag is added, pointing to HtmlJavascript

HtmlCss needs to be type: text/css
HtmlJavascript needs to be type: text/javascript

This provides the simplest solution to customizing the output of the HTML
representations. Some things to keep in mind:

* the existence of HtmlCss makes it be the _only_ css file. If
  profile.css is desired to be around, it should be @import-ed in HtmlCss
* If these tiddlers exist they are added to every html representation:
  tiddler, tiddlers, bag, bags, recipe, recipes and /profiles/<user>.
  This means that both CSS and Javascript needs to have some intelligence
  about what gets "turned on". You can't just assume one is on a
  friendly_uri page.

This puts a larger responsibility on the customizer, but it means that
the server side is simultaneously simple, flexible and powerful.