Optimize listing /bags and /recipes under ControlView
This works by avoiding the use of filters entirely, or listing
all the bags or recipes on the system. Instead it short circuits
directly to the list_entities method with the entities that
ought to be displayed.

Similar things might be possible for /search results, but we'll
cross that bridge at another time.