Add support for a /tiddlers URI
This presents a serialization of all the tiddlers in the correct
recipe for the current space, based on the URI and the membership
status of the current user. In otherwords is shows the tiddlers
either from the <space>_public recipe or the <space>_private recipe.

Normal content negotiation routines operate. The default representation
is the same as elsewhere: HTML

This is one of several "bad" URIs in tiddlyspace because it is
a different resource depending on context.