annotate tag definition description
since tags are special in that they are a list, a designer may
want to style them differently (e.g. put them on the same line)

currently it is very hard to do that with the current html. This provides
a route to do so, and establishes a "meta-<field>" class that could be used
on other elements if necessary.

I have not used the class tag here, to allow easy parsing of the page
by robots