Make the title of search results not be super ugly.
The title of an html representation of tiddlers that are a search
result comes from the query used to generate the results. When
that query is modified to optimize the ControlView process, the
result is a very ugly title.

So, instead of using the generated title, we save aside the original
query during ControlView and recreate the title before given it
to the results template.

I'm not particularly pleased with this solution. It is is yet one more
conditional behavior to adapt to the problems introduced by ControlView.

We have so many of those now.