Correct some what amount to typos in links.tid
The feeds we generate are _not_ RSS, they are Atom and while the
distinction is mostly meaningless to consumers, I don't want to be
called out by some pedant. So either "feed" or "Atom feed" is better.
I went with Atom feed.

Made the url and explanatory text on the tag search match.

We should add what I think is the most useful search + atom uri:

  http://tiddlyspace.com/search.atom?q=modifier:<someuser>

(that will provide the 20 most recently modified things from
<someuser>, whatever space it is from)

Hopefully this will notify @jdlrobson and @hugheth