Override tiddler_url to be more space conscious.
Update tiddlers serialization to use tiddler_url, instead
of generating url from scratch. This requires passing the
method and environ into the template, which is a bit weird,
but since we have decided that we control templates, is okay.
It is also most dry in this case.

Adjust the tiddlers serialization so that the tiddler_url
becomes aware if we are in a bag or recipe context. This
allows us to preserve that context when following links
from /tiddlers or /recipes/<foo>/tiddlers pages. This is
import for spaces/recipes which compose content by inclusion
and want to use space-wise html css and javascript (requiring
that we stay in the current context).

The overall win on this is that search results, in both atom
and html, will link to the result in its space, rather than
in the main server_host domain, providing a greater sense of
identity and probably improved SEO.

We'll have to watch this to see how it goes, it may require
some tweaking.

By introducing the overriding method we also open the possibility of
some more complex code which would link to friendly_uris
under the right circumstances, rather than bag/recipe uris. That's
not done in this commit as I think we need some time to work
out whether we want it or not.