remove references to images in jquery-ui.custom.css
note this is a short term fix
takenote currently uses this library and just the autocomplete css so much of
this theme may be redundant.

long term it would probably be better for takenote to work out which bits of the theme
it wants and remove this tiddler all together.