When /tiddlers is access as a wiki read ServerSettings
"as a wiki" means that the betaserialization or lazybetaserialization
is requested, via an Accept header or .wiki or .lwiki extensions.

We must only do the read on those serializations because reading
ServerSetting can change the query string and available filters,
which we don't want to do just any ol time.