move system-applications content into the tiddlyspace bag
after discussion with @jermolene @colmbritton and @bengillies
it was decided that system applications should not be optional since if
it is not, the backstage button links will 404 on spaces without it. As a result
it makes sense to add these to the tiddlyspace bag so that they cannot
be removed.