fix #775 if a SiteTitle (or similar) exists, don't try to create it when initing space.
The code assumes the user has never had a chance to write to their
space before the TiddlyWiki is created for the first time. With
the various "apps" this is no longer true.

The implementation just checks to see if a tiddler exists before
choosing to save it or not.