force backstage to check membership to current space
when no space is passed to the ts init function it reads the
current domain name for the current space (note where the domain
name is the same as the host in /status ts.js returns no space so
we don't need to worry about when this is run of the root)

when a space is passed, membership is checked and the body is given
a class showing the current state. The css rules make use of this
to grey/gray out the write button in the apps list to hint that it
is disabled but also allow a user to still navigate to it.