Establish a cache-manifest repudiator.
See: https://github.com/TiddlySpace/tiddlyspace/issues/781

If the request is a GET and the content-type of the resource
being sent is text/cache-manifest, then append a comment with
the tiddlyspace version in it.

The test for the repudiator just tests that it does the job, but
it doesn't actually confirm or deny that browsers will do the
right thing. It entirely depends on whether they send appropriate
cache-control headers (or lack thereof) when making requests to
check the manifest files. But now this is in place we can see how
to tweak things.

The reason for the size of this is commit being so large is that
lots of fiddling about was needed to make the version number
easily accessible to multiple files without creating circular
imports.

Luckily this also means that the __init__.py now makes considerable
more sense, and all the fixups can be in the fixups.py file, and the
plugin initialization in plugin.py.

This is just more tidy, thank you very much.