Add _cache-max-age info to the most common javascript files.
These choices were made based on the tiddlyspace.com access logs
indicating how often files are loaded or validated. Once these
have moved down the list, we can figure out others that can be
cached. At the moment this just gets javascript files (via .meta
files) and one html typed tid. reset.css and app.css are also high
on the list, but to get them we either need to support non-js meta
files in twp.twimport, or set the max-age field through some other
mechanism.

The time chosen is 1 hour. The assumption here is that we want the
caching to last during a user's browsing session, while still allowing
some flexibility for deployments and the like. Another assumption
is that for tiddlers we expect to be changing often, or to be needing
to change in an emergency, setting this field is not a good idea.