Reify the auxbags concept.
Each space gets a bag called <spacename>_auxbags[1], currently
editable by members[2] of the space, but not present in the recipes.

When ControlView and DropPrivs run they look at this bag, listing
the tiddlers within. The titles of the tiddlers are treated as bags
that "pass" for ControlView and DropPrivs[3].

If the tiddlers name a bag which does not exist, listing the /bags
of a space will fail[4], so it is expected that when a bag is deleted
the tiddler list should be updated.

This is a prototype implementation to get a sense of what's going on.
A more complete implementation would address some of the concerns
below.

See test/test_web_auxbags.py to see the stuff at play.

Please comment.

[1] This will require an update of existing spaces, as when the
archive bag was added.
[2] This is the main concern. How do we effectively control access?
Perhaps with a validator? Some magic needs to be included with the
request?
[3] This "pass" for both idempotent and not HTTP methods. The bag's
policy is expected to provide any desired protection. One possible
extension would be to be able to declare that an auxbag entry can
say that it should be "mounted" "read only", which would cause DropPrivs
to drop to GUEST for that request.
[4] We may wish to change this as you may choose to auxbag a bag which
is not under you control, but you still want in your space, and then
that bag may go away.