by default load links in new window
this feels very wrong. Sometimes you want it to be the parent
(e.g. i'm searching for one of my spaces and clicking on the link I
expect to navigate away from the current space)

Sometimes you want to load help in a certain window.

Obviously the user always has a choice with right click but they
don't have a choice to load in the parent window.

I'd be willing to revert this but going with @bengillies on this one