prevent excessive ajax requests in older versions of firefox
in firefox 5.0 the backstage triggers lots of /status requests until
it is opened putting unnecessary strain on the server

cdent discovered:
http://support.mozilla.com/en-US/questions/877508

it seems that the iframe will continually refresh until the backstage button
is clicked and the iframe displayed. using visibility instead of display solves
this. note because it is absolutely positioned no known side effects to how
page is rendered