allow SiteIcon uploading to _account page
When a SiteIcon hasn't been set there is a broken image in the universal
backstage next to the _account icon. It seems logical that if I click the
link I should be able to fix that.

This refactors the SiteIcon uploading code so both _account and _space
can use it. _account has additional logic to only allow SiteIcon uploading
from a homespace. If I navigate to fnd.tiddlyspace.com/_account at the moment
it allows me to change password etc.. but it will not be possible because
of controlview to upload a SiteIcon. In this situation we just hide it