Fixed following bubble handling of followers
This fixes bug #908.

The follow tiddlers macro now defaults to true if you don't specify
anything. You need to specify "follow:false" in the parameters to have
it report on everyone's tiddlers.

This also fixes a related bug that meant manually specifying a tiddler
to report on didn't work correctly. This was due to the scroll behaviour
checking for existence of the wrong tiddler (i.e. the tiddler being
reported on, not the tiddler the macro is in).