raise a HTTP409 when adding user to ANY or NONE policy
fix #910

This simply doesn't allow the action to continue, sending back the
error. This required because ANY or NONE policies have just ANY
or NONE as members. We can't guess what the intent of the current
requestor combined with the original policy author is, so it is
best not to guess.

This issues only comes up with spaces that have changed their default
policies (i.e. very few).