Extract the handling of the ServerSettings tiddler to middleware.
This means that it happens on every request, instead of just
some requests. The upshot of this is that now it can be used
in a more flexible fashion to set other thing. See

   http://cdent.tiddlyspace.com/tw5ikifier#d49c7e3599cb19a47aec9211f257beb6

for some discussion.

There's currently a large bug: query string additions are made
all the time, which is not desired, as it will cause filter strings,
for example, to always be applied, not just some of time.