Fix a long standing bug on the 404 page.
The 404 page is only supposed to present a link to create a tiddler
or a links to possible other tiddlers you might be looking for if
in fact you are looking at a tiddler URI. Instead it was doing it
pretty much all the time. I think this was @jdlrobson , so pinging
him so he's aware.

Anyway, the change says to only present the tiddler info if:

* the path is only one segment long and is not 'bags' or 'recipes'
* the path is 4 segments long