Temporarily disable the app switcher as default
This changes the default page for new spaces back to TiddlyWiki. This is
because the app switcher doesn't really give any direction or guidance
as to what users are supposed to do. In short, it's not ready yet.

This commit just comments out the code in the function (rather than
removing it) on the basis that the app switcher is a better long term
home page, but isn't the best one at the moment.