Added instructions about the reply button to the help text
The How to socialise tiddler now includes a brief text about replying to
tiddlers.

We had a debate about the correct terminology for this and whether
replying/responding was the correct way of phrasing what's happening.
I've tried to capture the fact that it's not necessarily an actual
"reply", but I'm not sure the text is quite there.