Fix issue #949 only show 1 login method
The previous method of doing this relied on the :target pseudo-selector,
which only works in certain browsers and so included a hack to always
show both.

As showing both was the thing we're trying to avoid, I've switched it to
a standard class based hide/show mechanism instead, which should work in
all browsers.