Add a 'friendly' flag to tiddler_url override.
This makes it possible to choose to present a tiddler's url
at its friendly location. If the space can't be determined,
then friendly is not possible, so it won't try.

This is primarily designed for use in places like the search.html
template where its probably best to link to tiddlers in the
context that people will share them. tiddlers.html should not
use it because that is linking to tiddlers _in_ recipes and bags
explicitly.