Further refinements to the search.html template
This version is somewhat more tidy in how it creates the
relevant links. The main change here is making the space_bag
conditional available in the template. This makes it possible
to say "are we working with something that should be presented in
a spacewise way, or not.

It would be useful if @pads and @fnd could comment. What remains to be
done includes:

* getting the ordering of the per tiddler info correct
* choosing whether the extra verbiage is useful valid
* choosing whether the semantics and containership of the html is correct
* deciding if we want to use some javascript to cook the modified
  dates (using something like timeago, as in http://tapas.tiddlyspace.com )
* styling (very boring and difficult to read right now)
* use of siteicons?

What's there is deployable, by virtue of being better than what
is currently out there, but it would be better to be even better.