fix #1001 make SiteInfo on space creation.
This is just tidier and avoid confusion from inclusion.

As @fnd points out a similar thing probably ought to be done
for the GettingStarted tiddler as it also can be included in
awkward ways. However, in that case, that is because space
authors are misusing/abusing the GettingStarted tiddler, which
ought to go away once starting has been gotten.