some jslint cleanups to a recently bugfixed tiddlywiki plugin
This certainly doesn't get all the issues, but tidies things
up a bit. Due to TiddlyWiki's use of lots of globals it's never
gonna be complete.

We should probably do this on most of the plugins. If we had the
problem that was fixed by 2f6413b73d76a529c15b22cc666bcc41491af39e
would have been found by linting.