Clarify and constrain test environment
I'm not sure when, but at some point the test_environment
lost track of where its templates were and stopped using
the correct database configuration when starting up.

These changes make sure that

* templates and the current code are linked into the test_instance
* the test_instance uses the right database settings
* the Session for the current test module is freshly configured
* that we don't recursively generate test_instances within each
  one

This is one of a few different instances of

* how did it ever work right before
* these tests are just wrong and we ought to make them right

This solution was reached by trial and error and was only discovered
because when manipulating templates for tests the changed templates
were not being reflected in the tests.