Fix links in profiles
fix #1038

This addresses to separate but related issues:

1. recently changed tiddler lists were non-friendly, non-space
   links, now they are
2. wikilinks in the profile text, when relative (markdown, twikifier)
   were going to 404 because they are looked for in tiddlyspace.com/profiles

Fix 2 was relatively straightforward: adding a <base> pointing
to the user's space. This gets behavior that is similar to what would
happen if the profile was viewed in the user's space.

Fix 1 was weird. Simply adding 'friendly=True' to the tiddler_url call
when listing tiddlers didn't work. This is because the fixups were
not applying in the profile module. Viewing the diff indicates the
necessary changes.

This change also adds a title to the profile page. Apparently this
had never been there.