Link app files with common space & clean correct common dir
The app files are part of tiddlyspace and should be symlinked.  The
files in the lib directory are copied over from an external source.

If these file are to be picked up automatically by devstore2 then the
developer should create the symlinks needed from the desired
repository to the src/common directory.