correctly call the callback in space reset
The previous code would only work when the number of private
tiddlers was one (1).

We want to compare the current index with the number of tiddlers,
and when we've reached the end, then do the callback.

This, however, is still not fully correct. All we are saying is
"the last DELETE has come back with success, so callback". It is
entirely possibly that some other, earlier, tiddler's DELETE is
still pending.

So really we probably want some kind of counter, but this is
at least better.

defering to @pads and @colmjude for final consideration