Get rid of the link to favicon.ico
It is redundant. If favicon.ico is in the recipe then it is at /
as well and that's where browsers except to find it by default
so there's no need for a special instruction, which doesn't appear
to work anyway (at least not in all browsers).

fixes #1061

The other entries seem okay. The bag as the source of an atom feed
is the most straightforward way to get a public feed (since the one
at /tiddlers would include tiddlers that come from other spaces
and things related to deployment which would be noisy).