Switch to beta only for "other" tiddlywiki releases
The current TiddlyWiki build process doesn't make alphas
at this time, so we don't want to confuse by using old
ones. This commit changes the code so it does the
right thing, but fails for the time being as the
generated externalized beta file points to the wrong files.

There are also no current test for an externalized stable
version, as that file does not yet exist (but hopefully will).

Tests associated with this process have been marked xfail,
and this is on a branch until resolution.

Part of issue #1060