Add the htmltemplate concept to ServerSettings
This provides a baseline for starting the
http://tsroadmap.tiddlyspace.com/HTML%20Fixification work
in an iterative fashion, on the deployed service, rather than
needing to get it all done at once.

The value of htmltemplate (defaul '') is added as a prefix
(+ '_') to the name of the template being loaded by the
html serialization.

For this to work the ServerSettings middleware must run before
ControlView, since ControlView uses the serialization directly.

The template loading routines have been adjusted so that HtmlCss
and HtmlJavascript are only loaded under two circumstances:

* no prefix is set (the current mainline behavior)
* the base template name is in a list of CUSTOMIZABLES
  (friendytiddler.html, friendlytiddlers.html, search.html)

The serialization is adjusted so that tiddler_as uses friendlytiddler.html
instead of tiddler.html when the uri is friendly. list_tiddlers
is changed similarly.