Associate a css tiddler with the htmltemplate prefix.
I'm not sure if this is the technique that will stay, but for
the time being it works.

Also adding in an initial structure for the fix1 templates, using
fix1_base.html. This is the template that the rest of the fix1 templates
will extend/subclass.

Unfortunately one of the thing I learned during this batch of work
is that the data being based from tiddler_as to the templates is
redundant and messy. To fix that I will need to change the default
templates, not just the new ones.

The next steps are to create very basic versions of the necessary
templates and then tune them, with help from the likes of @fnd, @colmjude,
@pads etc. I'm fairly agnostic about what the HTML ends up being,
I just know the current set is just not great.

The new set will be (the fix1 prefix is a handy stub):

customizables (HtmlCss and HtmlJavascript applied, as well as a nice
default css (from fix1.css), useful data for UI):

* fix1_friendlytiddler.html
* fix1_friendlytiddlers.html
* fix1_search.html

other (max data, limited or no css):

* fix1_bag.html
* fix1_bags.html
* fix1_recipe.html
* fix1_recipes.html
* fix1_tiddler.html <- non friendly, including revisions
* fix1_tiddlers.html <- non friendly , including revisions