add a macro for modified and created timestamps
A possible fix or start at fix for #1089

A cascade of activity led to this:

* tiddlyweb was changed to have an easy way of getting a datetime
  object from a tiddler.timestamp
* tiddlywebplugins.templates was changed to have a filter `rfc_3339`
  to create an isoformat timestamp, legal for the datetime attribute
  of the <time> element
* a macros.html template, containing macros was added to the templates
  here
* anywhere modified or created timestamps were shown now use the macro

The macro gives a class of 'modified' or 'created' to the time element
which may mean that some of the surround HTML should change.

Comments from @fnd, @pads and @colmjude desired. I reckon this is
a move in the right direction, but may need some more tweaking.

I'm hoping that the use of macros will make the construction of the
templates a bit more straightforward and consistent.