fix #1097 again
The previous fix put the sort after any existing filters which would
lead to bizarre result. Now the sort is inject before existing filters,
only if there is no existing sort filter.

For example if you stated limit=10 the result would be the 10 first
tiddlers from the database ordered tiddlers, sorted by -modified.

Now we get the top ten most recently modified tiddlers. This is
what we've decided we want but it does come at a cost: The entire
collection is traversed to do the default sort. An expected cost.