Add status.js to friendly representations.
fix #1099

Also fix title handling in clean1. This is redundant and not the
desired way of doing things, but want a quick fix for dealing with
reply handling. Fixes https://github.com/TiddlySpace/ts-reply/issues/8

To get this to work required that _reply-loader.js was updated to
not load status itself.